Fix projectile movement
Projectiles can get destroyed in Move(), so using a ForEach isn't
appropriate. Iterating over list of projectiles backwards is a better
(but not foolproof) idea.

There may still be an issue here, if two projectiles destroy eachother
in the same tick.

Applied same change to Enemy movement to be safe.